address various comments from code review
