+++ assignment-client/src/AssignmentClientChildData.cpp
@@ -1,0 +1,8 @@
+
+#include "AssignmentClientChildData.h"
+
+
+AssignmentClientChildData::AssignmentClientChildData(QString childType) :
+    _childType(childType)
+{
+}
